[FEAT] 카테고리 DTO 추가 및 카테고리 모델 카테고리 리스트 변수 추가

+++ feature/home/src/main/java/com/foodkeeper/feature/home/component/dialog/RecipeRecommendationDialog.kt
@@ -70,7 +70,7 @@ fun RecipeRecommendationDialog(
     onGenerateRecipe: (List<Food>) -> Unit
 ) {
     // 선택된 식재료 ID 목록
-    var selectedFoodIds by remember { mutableStateOf(setOf<Int>()) }
+    var selectedFoodIds by remember { mutableStateOf(setOf<Long>()) }
 
     Dialog(
         onDismissRequest = onDismiss,
